solar-panel-detection-main/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
pvlib==0.10.3
pandas==2.1.4
numpy==1.26.2
requests==2.31.0
google-cloud-bigquery==3.14.1
aiohttp==3.9.1
